Update terms of use
Latest updates to the English version from the legal team,
and translations to Indonesian, Thai, and Vietnamese.

Note that we don't have updated translations for Spanish
or Brazilian Portuguese at this time.

In the English version, I inserted the hyperlink to the
installed English version, so that it will be translated
properly in the future.  (For this time, I had to manually
insert the hyperlink in the Indonesian, Thai, and Vietnamese
translations.)

https://phabricator.endlessm.com/T15194