Change wording of Accept button in terms of use
We are changing the 'Accept and Continue' button to 'Accept' in the FBE.
The terms of service mention that you accept the terms by clicking
'Accept and Continue' so this unfortunately must be changed in the
terms document, in every language.

Some of the translations in the terms of service documents already
didn't match the wording in the UI; I suppose because a translator on
Transifex changed the translation in the UI without knowing that it had
to correspond to wording in a legal document. This problem potentially
still exists, so we should really reword the terms of service to not
specifically mention the language on the button.

- en: changed to 'Accept'
- es: changed to the PO file's translation of 'Accept'
- id: Translation in TOS ('menerima dan melanjutkan') did not match the
  translation in the PO file in the first place ('terima dan
  lanjutkan'). 'me(ng)-' is a prefix on verbs emphasizing that the
  speaker is doing the action so this is roughly like 'I accept and I
  continue' vs. 'Accept and continue' in English. I went with the word
  that matches how 'Accept' is translated upstream: 'terima'.
- pt_BR: changed to the PO file's translation of 'Accept'
- th: This was a bit iffy. I did not find the PO file's translation of
  'accept and continue' ('ยอมรับและไปต่อ') in the document. However, there
  was a likely looking phrase in quotes in about the right place in the
  document ('ยอมรับและดำเนินการต่อ') and sure enough Google translates that
  as 'accept and continue'. The phrase in the PO file translated to
  'accept and go' according to Google. Changed to 'ยอมรับ' which is
  'accept' according to both Google and the PO file.
- vi: Also a bit iffy. The document had 'đồng ý và tiếp tục' which
  Google translates as 'agree and continue', whereas the PO file has
  'Chấp nhận và tiếp tục' for 'accept and continue'. Changed to 'Chấp
  nhận' which is 'Accept' according to both the PO file and Google.

https://phabricator.endlessm.com/T25162