Convert to PDF using LibreOffice Flatpak
The terms of service are written in .docx format, and installed in .pdf
format. We commit the .pdfs to Git. Looking at the Git history (a205a1d)
it appears this is due to issues running unoconv in Jenkins. The
conversion is also not deterministic – better that unrelated updates in
this package do not cause the PDFs to be needlessly updated on users'
systems.

It seems reasonable to expect that most (all?) people updating the terms
of service PDFs will be using Endless. unoconv is not available on
Endless. Fortunately, LibreOffice's `--convert-to pdf` option works
fine, and LibreOffice is installed as a Flatpak on Endless OS. Add a
script to use that.

However... to run `make` you are likely using a toolbox container, but
the LibreOffice Flatpak is installed on the host system. Teach the
script to handle that case, too.